refactor: move user module under modules directory
