+++ apps/api/src/graphql/schema.ts
@@ -5,14 +5,19 @@ import {GraphQLSchema} from 'graphql';
 import path from 'node:path';
 
 const getTypeDefs = async () => {
-  return loadFiles(path.join(__dirname, '../**/graphql/*.graphql'));
+  return loadFiles(
+    path.join(__dirname, '../modules/**/graphql/typedefs/*.graphql'),
+  );
 };
 
 const getResolvers = async () => {
-  return loadFiles(path.join(__dirname, '../**/resolver/*.resolver.*'), {
-    ignoreIndex: false,
-    extensions: ['.js', '.ts'],
-  });
+  return loadFiles(
+    path.join(__dirname, '../modules/**/graphql/resolvers/*.resolver.*'),
+    {
+      ignoreIndex: false,
+      extensions: ['.js', '.ts'],
+    },
+  );
 };
 
 export const initializeSchema = async (): Promise<GraphQLSchema> => {
@@ -20,12 +25,10 @@ export const initializeSchema = async (): Promise<GraphQLSchema> => {
 
   const typeDefs = mergeTypeDefs(await getTypeDefs());
 
-  const graphqlSchema = makeExecutableSchema({
+  return makeExecutableSchema({
     typeDefs,
     resolvers,
     // https://stackoverflow.com/a/57218081/3783238
     inheritResolversFromInterfaces: true,
   });
-
-  return graphqlSchema;
 };
