+++ codegen.yml
@@ -1,4 +1,4 @@
-schema: 'apps/api/src/**/graphql/*.graphql'
+schema: 'apps/api/src/modules/**/graphql/typedefs/*.graphql'
 generates:
   apps/api/src/generated/graphql.generated.ts:
     config:
